lesson7.py

+++ lessons/lesson7.py
@@ -44,23 +44,51 @@ def get_users():
 #update_user(Ronaldo, 10)
 #get_users()
 
-def delete_user(rowid):   
-    cursor.execute(
-        "DELETE FROM users WHERE rowid = ?", 
-        (rowid,)
-    )
+#def delete_user(rowid):   
+    #cursor.execute(
+        #"DELETE FROM users WHERE rowid = ?", 
+        #(rowid,)
+    #)
+    #connect.commit()
+    #print("Deleted")
+
+#delete_user(31)
+#get_users()
+def update_users(ids, name=None, age=None, hobby=None):
+    if isinstance(ids, int):
+        ids = [ids]
+
+    for user_id in ids:
+        if name is not None:
+            cursor.execute("UPDATE users SET name = ? WHERE rowid = ?", (name, user_id))
+        if age is not None:
+            cursor.execute("UPDATE users SET age = ? WHERE rowid = ?", (age, user_id))
+        if hobby is not None:
+            cursor.execute("UPDATE users SET hobby = ? WHERE rowid = ?", (hobby, user_id))
+
     connect.commit()
-    print("Deleted")
+    print("Users updated")
 
-delete_user(31)
+def delete_users(ids):
+    if isinstance(ids, int):
+        ids = [ids]
+
+    for user_id in ids:
+        cursor.execute("DELETE FROM users WHERE rowid = ?", (user_id,))
+
+    connect.commit()
+    print("Users deleted")
+
+
+update_users([1, 2, 3], name="Alex")
+update_users(range(4, 7), hobby="Chess")
+update_users(8, age=30)
+
+delete_users([10, 11])
+delete_users(range(20, 25))
+delete_users(30)
+
 get_users()
-
-def delete_users():
-    cursor.execute("DELETE FROM users WHERE rowid = ?")
-    for i in range(11, 20):
-        delete_user(i)
-get_users()
-
